feat(project): moved crawler to one file
